+++ cvxpy/expressions/variables/__init__.py
@@ -17,7 +17,7 @@ You should have received a copy of the GNU General Public License
 along with CVXPY.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-from .bool_var import BoolVar
-from .int_var import IntVar
-from .variable import Variable
-from .semidefinite import Semidef
+from cvxpy.expressions.variables.bool_var import Bool
+from cvxpy.expressions.variables.int_var import Int
+from cvxpy.expressions.variables.variable import Variable
+from cvxpy.expressions.variables.semidef_var import Semidef